requirement.txt
numpy==1.24.3
pandas==2.0.3
scipy==1.11.3
librosa==0.10.2.post1
pydub==0.25.1
soundfile==0.12.1
scikit-learn==1.3.1
transformers==4.35.0
flask==3.0.0
fastapi==0.104.1
gunicorn==21.2.0
uvicorn==0.23.2
matplotlib==3.8.0
seaborn==0.13.0
pytest==7.4.3
docker==6.1.3
python-dotenv==1.0.1
tqdm==4.66.1
torch==2.0.0
torchaudio==2.0.1
sounddevice==0.4.6
